Create MULTITHREAD  PROGRAMMING
AIM : To write a java program that implements a multi-threaded program
which has three threads. First thread generates a random integer every 1
second. If the value is even, second thread computes the square of the
number and prints. If the value is odd the third thread will print the value
of cube of the number.